training_system/requirements.txt
Flask
python-docx
Werkzeug
Pillow
opencv-python
numpy
rembg
onnxruntime
lxml
openpyxl